+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+# [1.1.0](https://github.com/akhromi-ahmad/athena/compare/v1.0.0...v1.1.0) (2026-06-10)
+
+
+### Features
+
+* add beta prereleases for dev branch ([de3ae34](https://github.com/akhromi-ahmad/athena/commit/de3ae3462d6e8586985552147d86a0dc0c8cc290))
+* add simple login page, create UserController, update signup page navigation and web routes ([eae4c57](https://github.com/akhromi-ahmad/athena/commit/eae4c57b0d9a873bc53591baee5bdaaf76be2c14))
+* dev signup, grouping views folder ([14176f1](https://github.com/akhromi-ahmad/athena/commit/14176f1b06186c03418beaec8e9ef2717f304f21))
+* implement user registration and login flows with validation and tests ([6aa3336](https://github.com/akhromi-ahmad/athena/commit/6aa33365aae2918e5520721d9103f92ec852169c))
+* update user tbale & add the products table ([f428bcc](https://github.com/akhromi-ahmad/athena/commit/f428bcc3e930cac2713c05a59376c1c72505adf0))
+
+# [1.1.0-beta.2](https://github.com/akhromi-ahmad/athena/compare/v1.1.0-beta.1...v1.1.0-beta.2) (2026-06-10)
+
+
+### Features
+
+* implement user registration and login flows with validation and tests ([6aa3336](https://github.com/akhromi-ahmad/athena/commit/6aa33365aae2918e5520721d9103f92ec852169c))
+
+# [1.1.0-beta.1](https://github.com/akhromi-ahmad/athena/compare/v1.0.0...v1.1.0-beta.1) (2026-06-09)
+
+
+### Features
+
+* add beta prereleases for dev branch ([de3ae34](https://github.com/akhromi-ahmad/athena/commit/de3ae3462d6e8586985552147d86a0dc0c8cc290))
+* add simple login page, create UserController, update signup page navigation and web routes ([eae4c57](https://github.com/akhromi-ahmad/athena/commit/eae4c57b0d9a873bc53591baee5bdaaf76be2c14))
+* dev signup, grouping views folder ([14176f1](https://github.com/akhromi-ahmad/athena/commit/14176f1b06186c03418beaec8e9ef2717f304f21))
+* update user tbale & add the products table ([f428bcc](https://github.com/akhromi-ahmad/athena/commit/f428bcc3e930cac2713c05a59376c1c72505adf0))
+
+# [1.1.0-beta.1](https://github.com/akhromi-ahmad/athena/compare/v1.0.0...v1.1.0-beta.1) (2026-06-09)
+
+
+### Features
+
+* add beta prereleases for dev branch ([de3ae34](https://github.com/akhromi-ahmad/athena/commit/de3ae3462d6e8586985552147d86a0dc0c8cc290))
+* dev signup, grouping views folder ([14176f1](https://github.com/akhromi-ahmad/athena/commit/14176f1b06186c03418beaec8e9ef2717f304f21))
+* update user tbale & add the products table ([f428bcc](https://github.com/akhromi-ahmad/athena/commit/f428bcc3e930cac2713c05a59376c1c72505adf0))
+
 # [1.2.0-beta.2](https://github.com/akhromi-ahmad/athena/compare/v1.2.0-beta.1...v1.2.0-beta.2) (2026-05-31)